requirements.txt
zope.interface==4.3.2
attrs==17.4.0
py==1.5.3
pycodestyle==2.3.1
more_itertools==4.1.0
pluggy==0.6.0
setuptools==39.0.1
six==1.11.0
argcomplete==1.9.4
attr==0.3.1
colorama==0.3.9
funcsigs==1.0.2
numpy==1.14.3
Twisted==18.4.0
PyYAML==3.12
zope==4.0b4
